requirements.txt
wagtail==2.15.3
wagtailmenus==3.0.2
psycopg2==2.9.3
dj-database-url==0.5.0
django-storages==1.12.3
boto3==1.20.48
django-cors-headers==3.10.1
django-redis==5.2.0
redis==4.1.2
wagtail-robots==0.3.1
sentry-sdk==1.5.4
pymemcache==3.5.0
anyascii==0.3.0
asgiref==3.4.1
beautifulsoup4==4.9.3
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
django-cogwheels==0.3
django-filter==21.1
django-modelcluster==5.2
django-taggit==1.5.1
django-treebeard==4.5.1
djangorestframework==3.13.1
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==3.3
l18n==2021.3
openpyxl==3.0.9
Pillow==8.4.0
pytz==2021.3
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.1.0
telepath==0.2
typing_extensions==4.1.1
urllib3==1.26.8
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==3.0.2
xlwt==1.3.0
django_storage_url
